+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/GetNotificationSettingsUseCase.kt
@@ -1,4 +1,4 @@
-package io.github.kei_1111.withmo.core.domain.usecase.notification
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.NotificationSettings
@@ -6,6 +6,10 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
+
+interface GetNotificationSettingsUseCase {
+    operator fun invoke(): Flow<NotificationSettings>
+}
 
 class GetNotificationSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
