+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/GetSideButtonSettingsUseCase.kt
@@ -1,9 +1,15 @@
-package io.github.kei_1111.withmo.core.domain.usecase.side_button
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
+import io.github.kei_1111.withmo.core.model.user_settings.SideButtonSettings
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
+
+interface GetSideButtonSettingsUseCase {
+    operator fun invoke(): Flow<SideButtonSettings>
+}
 
 class GetSideButtonSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
