+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/GetSortSettingsUseCase.kt
@@ -1,15 +1,21 @@
-package io.github.kei_1111.withmo.core.domain.usecase.clock
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
+import io.github.kei_1111.withmo.core.model.user_settings.SortSettings
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
-class GetClockSettingsUseCaseImpl @Inject constructor(
+interface GetSortSettingsUseCase {
+    operator fun invoke(): Flow<SortSettings>
+}
+
+class GetSortSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
-) : GetClockSettingsUseCase {
+) : GetSortSettingsUseCase {
     override operator fun invoke() =
         userSettingsRepository.userSettings
-            .map { it.clockSettings }
+            .map { it.sortSettings }
             .distinctUntilChanged()
 }
