+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/GetUserSettingsUseCase.kt
@@ -1,8 +1,16 @@
 package io.github.kei_1111.withmo.core.domain.usecase
 
+import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.UserSettings
 import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
 
 interface GetUserSettingsUseCase {
     operator fun invoke(): Flow<UserSettings>
+}
+
+class GetUserSettingsUseCaseImpl @Inject constructor(
+    private val userSettingsRepository: UserSettingsRepository,
+) : GetUserSettingsUseCase {
+    override operator fun invoke() = userSettingsRepository.userSettings
 }
