+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/SaveClockSettingsUseCase.kt
@@ -1,8 +1,12 @@
-package io.github.kei_1111.withmo.core.domain.usecase.clock
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.ClockSettings
 import javax.inject.Inject
+
+interface SaveClockSettingsUseCase {
+    suspend operator fun invoke(clockSettings: ClockSettings)
+}
 
 class SaveClockSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
