+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/SaveModelFilePathUseCase.kt
@@ -1,8 +1,12 @@
-package io.github.kei_1111.withmo.core.domain.usecase.model_file_path
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.ModelFilePath
 import javax.inject.Inject
+
+interface SaveModelFilePathUseCase {
+    suspend operator fun invoke(modelFilePath: ModelFilePath)
+}
 
 class SaveModelFilePathUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
