+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/SaveNotificationSettingsUseCase.kt
@@ -1,8 +1,12 @@
-package io.github.kei_1111.withmo.core.domain.usecase.notification
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.NotificationSettings
 import javax.inject.Inject
+
+interface SaveNotificationSettingsUseCase {
+    suspend operator fun invoke(notificationSettings: NotificationSettings)
+}
 
 class SaveNotificationSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
