+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/SaveSortSettingsUseCase.kt
@@ -1,8 +1,12 @@
-package io.github.kei_1111.withmo.core.domain.usecase.sort
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.SortSettings
 import javax.inject.Inject
+
+interface SaveSortSettingsUseCase {
+    suspend operator fun invoke(sortSettings: SortSettings)
+}
 
 class SaveSortSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
