+++ core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/SaveThemeSettingsUseCase.kt
@@ -1,8 +1,12 @@
-package io.github.kei_1111.withmo.core.domain.usecase.theme
+package io.github.kei_1111.withmo.core.domain.usecase
 
 import io.github.kei_1111.withmo.core.domain.repository.UserSettingsRepository
 import io.github.kei_1111.withmo.core.model.user_settings.ThemeSettings
 import javax.inject.Inject
+
+interface SaveThemeSettingsUseCase {
+    suspend operator fun invoke(themeSettings: ThemeSettings)
+}
 
 class SaveThemeSettingsUseCaseImpl @Inject constructor(
     private val userSettingsRepository: UserSettingsRepository,
