--- core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/app_icon/GetAppIconSettingsUseCase.kt
@@ -1,8 +1,0 @@
-package io.github.kei_1111.withmo.core.domain.usecase.app_icon
-
-import io.github.kei_1111.withmo.core.model.user_settings.AppIconSettings
-import kotlinx.coroutines.flow.Flow
-
-interface GetAppIconSettingsUseCase {
-    operator fun invoke(): Flow<AppIconSettings>
-}
