--- core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/app_icon/SaveAppIconSettingsUseCase.kt
@@ -1,7 +1,0 @@
-package io.github.kei_1111.withmo.core.domain.usecase.app_icon
-
-import io.github.kei_1111.withmo.core.model.user_settings.AppIconSettings
-
-interface SaveAppIconSettingsUseCase {
-    suspend operator fun invoke(appIconSettings: AppIconSettings)
-}
