--- core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/model_file_path/GetModelFilePathUseCase.kt
@@ -1,8 +1,0 @@
-package io.github.kei_1111.withmo.core.domain.usecase.model_file_path
-
-import io.github.kei_1111.withmo.core.model.user_settings.ModelFilePath
-import kotlinx.coroutines.flow.Flow
-
-interface GetModelFilePathUseCase {
-    operator fun invoke(): Flow<ModelFilePath>
-}
