--- core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/theme/GetThemeSettingsUseCase.kt
@@ -1,8 +1,0 @@
-package io.github.kei_1111.withmo.core.domain.usecase.theme
-
-import io.github.kei_1111.withmo.core.model.user_settings.ThemeSettings
-import kotlinx.coroutines.flow.Flow
-
-interface GetThemeSettingsUseCase {
-    operator fun invoke(): Flow<ThemeSettings>
-}
