--- core/domain/src/main/kotlin/io/github/kei_1111/withmo/core/domain/usecase/theme/SaveThemeSettingsUseCase.kt
@@ -1,7 +1,0 @@
-package io.github.kei_1111.withmo.core.domain.usecase.theme
-
-import io.github.kei_1111.withmo.core.model.user_settings.ThemeSettings
-
-interface SaveThemeSettingsUseCase {
-    suspend operator fun invoke(themeSettings: ThemeSettings)
-}
