+++ feature/home/src/main/kotlin/io/github/kei_1111/withmo/feature/home/HomeViewModel.kt
@@ -16,7 +16,7 @@ import io.github.kei_1111.withmo.core.domain.repository.AppInfoRepository
 import io.github.kei_1111.withmo.core.domain.repository.OneTimeEventRepository
 import io.github.kei_1111.withmo.core.domain.repository.WidgetInfoRepository
 import io.github.kei_1111.withmo.core.domain.usecase.GetUserSettingsUseCase
-import io.github.kei_1111.withmo.core.domain.usecase.model_file_path.SaveModelFilePathUseCase
+import io.github.kei_1111.withmo.core.domain.usecase.SaveModelFilePathUseCase
 import io.github.kei_1111.withmo.core.featurebase.BaseViewModel
 import io.github.kei_1111.withmo.core.model.AppInfo
 import io.github.kei_1111.withmo.core.model.WidgetInfo
