+++ feature/setting/src/main/kotlin/io/github/kei_1111/withmo/feature/setting/sort/SortSettingsViewModel.kt
@@ -3,8 +3,8 @@ package io.github.kei_1111.withmo.feature.setting.sort
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.github.kei_1111.withmo.core.domain.usecase.sort.GetSortSettingsUseCase
-import io.github.kei_1111.withmo.core.domain.usecase.sort.SaveSortSettingsUseCase
+import io.github.kei_1111.withmo.core.domain.usecase.GetSortSettingsUseCase
+import io.github.kei_1111.withmo.core.domain.usecase.SaveSortSettingsUseCase
 import io.github.kei_1111.withmo.core.featurebase.BaseViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
