+++ feature/setting/src/main/kotlin/io/github/kei_1111/withmo/feature/setting/theme/ThemeSettingsViewModel.kt
@@ -3,8 +3,8 @@ package io.github.kei_1111.withmo.feature.setting.theme
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.github.kei_1111.withmo.core.domain.usecase.theme.GetThemeSettingsUseCase
-import io.github.kei_1111.withmo.core.domain.usecase.theme.SaveThemeSettingsUseCase
+import io.github.kei_1111.withmo.core.domain.usecase.GetThemeSettingsUseCase
+import io.github.kei_1111.withmo.core.domain.usecase.SaveThemeSettingsUseCase
 import io.github.kei_1111.withmo.core.featurebase.BaseViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
